requirements.txt
discord.py[voice]==2.4.0
motor==3.6.0
python-dotenv==1.0.0
psutil==6.0.0
# numpy==1.24.2
# pandas==1.4.3
# faiss_cpu==1.7.3
# torch==2.0.0
# torchvision==0.15.1
# timm==0.6.13
# tqdm==4.65.0
pillow==10.4.0
python-Levenshtein==0.23.0
aiohttp==3.9.5
yt-dlp==2024.9.27